+++ include/endstone/metrics/simple_bar_chart.h
@@ -34,17 +34,17 @@ public:
     {
     }
 
-    [[nodiscard]] std::optional<nlohmann::json::object_t> getChartData() override
+    [[nodiscard]] std::optional<JsonObject> getChartData() override
     {
         const auto map_values = get_values_();
         if (!map_values || map_values->empty()) {
             return std::nullopt;
         }
-        nlohmann::json::object_t values;
+        JsonObject values;
         for (const auto &[key, value] : *map_values) {
-            values[key] = nlohmann::json::array({value});
+            values[key] = JsonArray{value};
         }
-        return nlohmann::json::object_t{{"values", std::move(values)}};
+        return JsonObject{{"values", std::move(values)}};
     }
 
 private: